+++ lib/src/i18next.dart
@@ -1,6 +1,5 @@
+import 'dart:convert';
 import 'dart:ui';
-
-import 'package:flutter/foundation.dart';
 
 import 'options.dart';
 import 'utils.dart';
@@ -75,7 +74,35 @@ class I18Next {
     Locale locale,
   }) {
     assert(key != null);
-    locale ??= this.locale;
+
+    return _Translation(
+          locale ?? this.locale,
+          dataSource,
+          options,
+        ).t(
+          key,
+          context: context,
+          count: count,
+          variables: variables,
+        ) ??
+        key;
+  }
+}
+
+class _Translation {
+  _Translation(this.locale, this.dataSource, this.options);
+
+  final Locale locale;
+  final LocalizationDataSource dataSource;
+  final I18NextOptions options;
+
+  String t(
+    String key, {
+    String context,
+    int count,
+    Map<String, Object> variables,
+  }) {
+    assert(key != null);
 
     String namespace = '', keyPath = key;
     final match = RegExp(':').firstMatch(key);
@@ -84,35 +111,27 @@ class I18Next {
       keyPath = key.substring(match.end);
     }
 
-    return translateKey(namespace, keyPath,
-            context: context,
-            count: count,
-            variables: variables,
-            locale: locale,
-            options: options,
-            dataSource: dataSource) ??
+    return translateKey(
+          namespace,
+          keyPath,
+          context: context,
+          count: count,
+          variables: variables,
+        ) ??
         key;
   }
 
-  static String translateKey(
+  String translateKey(
     String namespace,
     String key, {
     String context,
     int count,
     Map<String, Object> variables,
-    Locale locale,
-    I18NextOptions options,
-    // TODO: remove this later
-    @required LocalizationDataSource dataSource,
   }) {
     if (context != null && context.isNotEmpty) {
       final contextKey = '${key}_$context';
       final value = translateKey(namespace, contextKey,
-          count: count,
-          variables: variables,
-          locale: locale,
-          options: options,
-          dataSource: dataSource);
+          count: count, variables: variables);
       if (value != null) return value;
     }
 
@@ -121,23 +140,16 @@ class I18Next {
       variablesWithCount['count'] ??= count;
 
       final pluralKey = pluralize(key, options.pluralSuffix, count, locale);
-      final value = translateKey(namespace, pluralKey,
-          variables: variablesWithCount,
-          locale: locale,
-          options: options,
-          dataSource: dataSource);
+      final value =
+          translateKey(namespace, pluralKey, variables: variablesWithCount);
       if (value != null) return value;
     }
 
-    return find(namespace, key,
-        variables: variables,
-        locale: locale,
-        options: options,
-        dataSource: dataSource);
+    return find(namespace, key, variables: variables);
   }
 
   /// Returns the pluralized form for the [key] based on [locale] and [count].
-  static String pluralize(String key, String suffix, int count, Locale locale) {
+  String pluralize(String key, String suffix, int count, Locale locale) {
     if (count != 1) {
       final number = _numberForLocale(count.abs(), locale);
       if (number >= 0)
@@ -148,7 +160,7 @@ class I18Next {
     return key;
   }
 
-  static int _numberForLocale(int count, Locale locale) {
+  int _numberForLocale(int count, Locale locale) {
     // TODO: add locale based rules
     return -1;
   }
@@ -156,7 +168,7 @@ class I18Next {
   /// Given a key with multiple split points (`.`), this method navigates
   /// through the objects and returns the last node, expecting it to be a
   /// [String], null otherwise.
-  static String evaluate(String path, Map<String, Object> data) {
+  String evaluate(String path, Map<String, Object> data) {
     final keys = path.split('.');
 
     dynamic object = data;
@@ -171,15 +183,7 @@ class I18Next {
     return object;
   }
 
-  static String find(
-    String namespace,
-    String key, {
-    Map<String, Object> variables,
-    Locale locale,
-    I18NextOptions options,
-    // TODO: remove this later
-    @required LocalizationDataSource dataSource,
-  }) {
+  String find(String namespace, String key, {Map<String, Object> variables}) {
     // TODO: find out if namespaces are loaded early
     Map<String, Object> data = dataSource(namespace, locale);
     final value = evaluate(key, data);
@@ -191,29 +195,26 @@ class I18Next {
 
     String result;
     if (value != null) {
-      result = interpolate(value,
-          variables: variables, locale: locale, options: options);
-      result =
-          nest(result, variables: variables, locale: locale, options: options);
+      variables ??= {};
+      result = interpolate(value, variables: variables);
+      result = nest(result, variables: variables);
     }
 
     return result;
   }
 
-  /// Replaces occurrences of matches in [target] for the named values
+  /// Replaces occurrences of matches in [string] for the named values
   /// in [variables] (if they exist), by first passing through the
-  /// [InterpolationOptions.formatter] before joining the resulting string.
+  /// [I18NextOptions.formatter] before joining the resulting string.
   ///
   /// - 'Hello {{name}}' + {name: 'World'} -> 'Hello World'.
   ///   This example illustrates a simple interpolation.
   /// - 'Now is {{date, dd/MM}}' + {date: DateTime.now()} -> 'Now is 23/09'.
-  ///   In this example, [InterpolationOptions.formatter] must be able to
+  ///   In this example, [I18NextOptions.formatter] must be able to
   ///   properly format the date.
-  static String interpolate(
+  String interpolate(
     String string, {
     Map<String, Object> variables,
-    Locale locale,
-    I18NextOptions options,
   }) {
     return string.splitMapJoin(
       options.interpolationPattern,
@@ -232,13 +233,45 @@ class I18Next {
     );
   }
 
-  static String nest(
-    String string, {
-    Map<String, Object> variables,
-    Locale locale,
-    I18NextOptions options,
-  }) {
-    // TODO: implement
-    return string;
+  /// Replaces occurrences of nested key-values in [string] for other
+  /// key-values. Essentially calls [I18Next.t] with the nested value.
+  ///
+  /// E.g.:
+  /// ```json
+  /// {
+  ///   key1: "Hello $t(key2)!"
+  ///   key2: "World"
+  /// }
+  /// i18Next.t('key1') // "Hello World!"
+  /// ```
+  String nest(String string, {Map<String, Object> variables}) {
+    return string.splitMapJoin(options.nestingPattern, onMatch: (match) {
+      RegExpMatch regExpMatch = match;
+      final key = regExpMatch.namedGroup('key');
+
+      String result;
+      if (key != null) {
+        final varsString = regExpMatch.namedGroup('variables');
+        String context;
+        int count;
+
+        final copy = Map<String, Object>.from(variables);
+        if (varsString != null && varsString.isNotEmpty) {
+          try {
+            final Map<String, Object> vars = jsonDecode(varsString);
+            if (vars != null) {
+              context = vars['context'];
+              count = vars['count'];
+              copy.addAll(vars);
+            }
+          } catch (error) {
+            assert(true, error);
+          }
+        }
+
+        result = t(key, context: context, count: count, variables: copy);
+      }
+      return result ?? regExpMatch.group(0);
+    });
   }
 }